refactor: split storage/mod.rs (2622 lines) into focused submodules
Behavior-preserving split of the storage stack boundary, which had grown
into a single 2.6k-line file mixing the trait API surface with every
backend's trait implementations. No logic changed — only module location,
use headers, and visibility (private -> pub(super) where an item now
crosses the new module boundary).

  storage/mod.rs            880  trait/type API surface + StorageStack
  storage/providers.rs      245  backend-selection provider wrappers
  storage/backends/mod.rs    28  re-export hub for the submodules
  storage/backends/convert.rs        191  Postgres->Turso row converters
  storage/backends/metadata_stores.rs   502  policy/observe/evolution impls
  storage/backends/lifecycle_stores.rs  384  blob/decision/wasm/ots impls
  storage/backends/query_trajectory.rs  459  query-plane/trajectory impls

Every external `storage::X` path still resolves unchanged (the public API
items remain pub at crate::storage). All 43 impl blocks are identical to
their pre-split form. cargo check --all-targets is warning-free; the
temper-server suite passes.

DST review: PASS (mechanical move, no determinism-relevant change).
Code review: PASS (verified behavior-preserving).
