+++ crates/temper-server/src/storage/backends/query_trajectory.rs
@@ -1,0 +1,459 @@
+//! Query-plane, data-only-create, and trajectory-sink store impls.
+
+use super::*;
+
+#[async_trait::async_trait]
+impl QueryPlaneStore for PostgresEventStore {
+    async fn upsert_projection(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_id: &str,
+        status: &str,
+        fields: &serde_json::Value,
+        sequence_nr: u64,
+    ) -> Result<(), PersistenceError> {
+        self.upsert_query_projection(tenant, entity_type, entity_id, status, fields, sequence_nr)
+            .await
+    }
+
+    async fn remove_projection(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_id: &str,
+    ) -> Result<(), PersistenceError> {
+        self.remove_query_projection(tenant, entity_type, entity_id)
+            .await
+    }
+
+    async fn query_field_index(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        where_clause: &str,
+        params: Vec<String>,
+    ) -> Result<Option<Vec<String>>, PersistenceError> {
+        PostgresEventStore::query_field_index(self, tenant, entity_type, where_clause, params)
+            .await
+            .map(Some)
+    }
+
+    async fn load_projection_fields_many(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_ids: &[String],
+        field_names: &[&str],
+    ) -> Result<Option<Vec<QueryProjectionFieldsRow>>, PersistenceError> {
+        self.load_query_projection_fields_many(tenant, entity_type, entity_ids, field_names)
+            .await
+            .map(|rows| {
+                Some(
+                    rows.into_iter()
+                        .map(|row| QueryProjectionFieldsRow {
+                            entity_id: row.entity_id,
+                            status: row.status,
+                            fields: row.fields,
+                        })
+                        .collect(),
+                )
+            })
+    }
+
+    async fn load_entity_catalog_rows(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_ids: &[String],
+    ) -> Result<Option<Vec<EntityCatalogRow>>, PersistenceError> {
+        self.load_entity_catalog_rows_pg(tenant, entity_type, entity_ids)
+            .await
+            .map(|rows| {
+                Some(
+                    rows.into_iter()
+                        .map(|row| EntityCatalogRow {
+                            entity_id: row.entity_id,
+                            status: row.status,
+                            fields: row.fields,
+                            sequence_nr: row.sequence_nr,
+                        })
+                        .collect(),
+                )
+            })
+    }
+
+    async fn projected_entity_counts_by_tenant(
+        &self,
+    ) -> Result<Option<Vec<(String, u64)>>, PersistenceError> {
+        PostgresEventStore::projected_entity_counts_by_tenant(self)
+            .await
+            .map(Some)
+    }
+}
+
+#[async_trait::async_trait]
+impl QueryPlaneStore for TursoEventStore {
+    async fn upsert_projection(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_id: &str,
+        status: &str,
+        fields: &serde_json::Value,
+        sequence_nr: u64,
+    ) -> Result<(), PersistenceError> {
+        self.upsert_query_projection(tenant, entity_type, entity_id, status, fields, sequence_nr)
+            .await
+    }
+
+    async fn remove_projection(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_id: &str,
+    ) -> Result<(), PersistenceError> {
+        self.remove_query_projection(tenant, entity_type, entity_id)
+            .await
+    }
+
+    async fn query_field_index(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        where_clause: &str,
+        params: Vec<String>,
+    ) -> Result<Option<Vec<String>>, PersistenceError> {
+        TursoEventStore::query_field_index(self, tenant, entity_type, where_clause, params)
+            .await
+            .map(Some)
+    }
+
+    async fn load_projection_fields_many(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_ids: &[String],
+        field_names: &[&str],
+    ) -> Result<Option<Vec<QueryProjectionFieldsRow>>, PersistenceError> {
+        self.load_query_projection_fields_many(tenant, entity_type, entity_ids, field_names)
+            .await
+            .map(|rows| {
+                Some(
+                    rows.into_iter()
+                        .map(|row| QueryProjectionFieldsRow {
+                            entity_id: row.entity_id,
+                            status: row.status,
+                            fields: row.fields,
+                        })
+                        .collect(),
+                )
+            })
+    }
+
+    async fn load_entity_catalog_rows(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_ids: &[String],
+    ) -> Result<Option<Vec<EntityCatalogRow>>, PersistenceError> {
+        TursoEventStore::load_entity_catalog_rows(self, tenant, entity_type, entity_ids)
+            .await
+            .map(|rows| {
+                Some(
+                    rows.into_iter()
+                        .map(|row| EntityCatalogRow {
+                            entity_id: row.entity_id,
+                            status: row.status,
+                            fields: row.fields,
+                            sequence_nr: row.sequence_nr,
+                        })
+                        .collect(),
+                )
+            })
+    }
+
+    async fn projected_entity_counts_by_tenant(
+        &self,
+    ) -> Result<Option<Vec<(String, u64)>>, PersistenceError> {
+        TursoEventStore::projected_entity_counts_by_tenant(self)
+            .await
+            .map(Some)
+    }
+}
+
+#[async_trait::async_trait]
+impl DataOnlyCreateStore for PostgresEventStore {
+    async fn create_data_only_entity(
+        &self,
+        record: DataOnlyCreateRecord<'_>,
+    ) -> Result<u64, PersistenceError> {
+        self.create_data_only_entity_native(
+            record.tenant,
+            record.entity_type,
+            record.entity_id,
+            record.status,
+            record.fields,
+            record.event,
+        )
+        .await
+    }
+}
+
+#[async_trait::async_trait]
+impl QueryPlaneStore for TenantStoreRouter {
+    async fn upsert_projection(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_id: &str,
+        status: &str,
+        fields: &serde_json::Value,
+        sequence_nr: u64,
+    ) -> Result<(), PersistenceError> {
+        let store = self.store_for_tenant(tenant).await?;
+        store
+            .upsert_query_projection(tenant, entity_type, entity_id, status, fields, sequence_nr)
+            .await
+    }
+
+    async fn remove_projection(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_id: &str,
+    ) -> Result<(), PersistenceError> {
+        let store = self.store_for_tenant(tenant).await?;
+        store
+            .remove_query_projection(tenant, entity_type, entity_id)
+            .await
+    }
+
+    async fn query_field_index(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        where_clause: &str,
+        params: Vec<String>,
+    ) -> Result<Option<Vec<String>>, PersistenceError> {
+        let store = self.store_for_tenant(tenant).await?;
+        TursoEventStore::query_field_index(&store, tenant, entity_type, where_clause, params)
+            .await
+            .map(Some)
+    }
+
+    async fn load_projection_fields_many(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_ids: &[String],
+        field_names: &[&str],
+    ) -> Result<Option<Vec<QueryProjectionFieldsRow>>, PersistenceError> {
+        let store = self.store_for_tenant(tenant).await?;
+        store
+            .load_query_projection_fields_many(tenant, entity_type, entity_ids, field_names)
+            .await
+            .map(|rows| {
+                Some(
+                    rows.into_iter()
+                        .map(|row| QueryProjectionFieldsRow {
+                            entity_id: row.entity_id,
+                            status: row.status,
+                            fields: row.fields,
+                        })
+                        .collect(),
+                )
+            })
+    }
+
+    async fn load_entity_catalog_rows(
+        &self,
+        tenant: &str,
+        entity_type: &str,
+        entity_ids: &[String],
+    ) -> Result<Option<Vec<EntityCatalogRow>>, PersistenceError> {
+        let store = self.store_for_tenant(tenant).await?;
+        TursoEventStore::load_entity_catalog_rows(&store, tenant, entity_type, entity_ids)
+            .await
+            .map(|rows| {
+                Some(
+                    rows.into_iter()
+                        .map(|row| EntityCatalogRow {
+                            entity_id: row.entity_id,
+                            status: row.status,
+                            fields: row.fields,
+                            sequence_nr: row.sequence_nr,
+                        })
+                        .collect(),
+                )
+            })
+    }
+
+    async fn projected_entity_counts_by_tenant(
+        &self,
+    ) -> Result<Option<Vec<(String, u64)>>, PersistenceError> {
+        let mut counts = Vec::new();
+        for tenant_id in self.connected_tenants().await {
+            let store = self.store_for_tenant(&tenant_id).await?;
+            if let Some((_, count)) = TursoEventStore::projected_entity_counts_by_tenant(&store)
+                .await?
+                .into_iter()
+                .find(|(tenant, _)| tenant == &tenant_id)
+            {
+                counts.push((tenant_id, count));
+            }
+        }
+        Ok(Some(counts))
+    }
+}
+
+pub(super) fn trajectory_source_label(source: &TrajectorySource) -> &'static str {
+    match source {
+        TrajectorySource::Entity => "Entity",
+        TrajectorySource::Platform => "Platform",
+        TrajectorySource::Authz => "Authz",
+    }
+}
+
+pub(super) fn trajectory_request_body_json(entry: &TrajectoryEntry) -> Option<String> {
+    entry.request_body.as_ref().and_then(|value| {
+        let serialized = serde_json::to_string(value).ok()?;
+        Some(if serialized.len() > 4096 {
+            let mut end = 4096;
+            while !serialized.is_char_boundary(end) {
+                end -= 1;
+            }
+            serialized[..end].to_string()
+        } else {
+            serialized
+        })
+    })
+}
+
+pub(super) fn trajectory_matched_policy_ids_json(entry: &TrajectoryEntry) -> Option<String> {
+    entry
+        .matched_policy_ids
+        .as_ref()
+        .and_then(|ids| serde_json::to_string(ids).ok())
+}
+
+#[async_trait::async_trait]
+impl TrajectorySink for PostgresEventStore {
+    async fn persist_trajectory_entry(&self, entry: &TrajectoryEntry) -> Result<(), String> {
+        let matched_policy_ids_json = trajectory_matched_policy_ids_json(entry);
+        let request_body_json = trajectory_request_body_json(entry);
+        let source = entry.source.as_ref().map(trajectory_source_label);
+
+        self.persist_trajectory(PostgresTrajectoryInsert {
+            tenant: &entry.tenant,
+            entity_type: &entry.entity_type,
+            entity_id: &entry.entity_id,
+            action: &entry.action,
+            success: entry.success,
+            from_status: entry.from_status.as_deref(),
+            to_status: entry.to_status.as_deref(),
+            error: entry.error.as_deref(),
+            agent_id: entry.agent_id.as_deref(),
+            session_id: entry.session_id.as_deref(),
+            authz_denied: entry.authz_denied,
+            denied_resource: entry.denied_resource.as_deref(),
+            denied_module: entry.denied_module.as_deref(),
+            source,
+            spec_governed: entry.spec_governed,
+            created_at: &entry.timestamp,
+            request_body: request_body_json.as_deref(),
+            intent: entry.intent.as_deref(),
+            matched_policy_ids: matched_policy_ids_json.as_deref(),
+        })
+        .await
+        .map_err(|e| {
+            format!(
+                "failed to persist trajectory entry for {}/{}/{} action {} in postgres: {e}",
+                entry.tenant, entry.entity_type, entry.entity_id, entry.action
+            )
+        })
+    }
+}
+
+#[async_trait::async_trait]
+impl TrajectorySink for TursoEventStore {
+    async fn persist_trajectory_entry(&self, entry: &TrajectoryEntry) -> Result<(), String> {
+        let matched_policy_ids_json = trajectory_matched_policy_ids_json(entry);
+        let request_body_json = trajectory_request_body_json(entry);
+        let source = entry.source.as_ref().map(trajectory_source_label);
+
+        self.persist_trajectory(TursoTrajectoryInsert {
+            tenant: &entry.tenant,
+            entity_type: &entry.entity_type,
+            entity_id: &entry.entity_id,
+            action: &entry.action,
+            success: entry.success,
+            from_status: entry.from_status.as_deref(),
+            to_status: entry.to_status.as_deref(),
+            error: entry.error.as_deref(),
+            agent_id: entry.agent_id.as_deref(),
+            session_id: entry.session_id.as_deref(),
+            authz_denied: entry.authz_denied,
+            denied_resource: entry.denied_resource.as_deref(),
+            denied_module: entry.denied_module.as_deref(),
+            source,
+            spec_governed: entry.spec_governed,
+            created_at: &entry.timestamp,
+            request_body: request_body_json.as_deref(),
+            intent: entry.intent.as_deref(),
+            matched_policy_ids: matched_policy_ids_json.as_deref(),
+        })
+        .await
+        .map_err(|e| {
+            format!(
+                "failed to persist trajectory entry for {}/{}/{} action {} in turso: {e}",
+                entry.tenant, entry.entity_type, entry.entity_id, entry.action
+            )
+        })
+    }
+}
+
+#[async_trait::async_trait]
+impl TrajectorySink for TenantStoreRouter {
+    async fn persist_trajectory_entry(&self, entry: &TrajectoryEntry) -> Result<(), String> {
+        let store = self.store_for_tenant(&entry.tenant).await.map_err(|e| {
+            format!(
+                "failed to resolve tenant store for trajectory entry {}/{}/{} action {}: {e}",
+                entry.tenant, entry.entity_type, entry.entity_id, entry.action
+            )
+        })?;
+        let matched_policy_ids_json = trajectory_matched_policy_ids_json(entry);
+        let request_body_json = trajectory_request_body_json(entry);
+        let source = entry.source.as_ref().map(trajectory_source_label);
+
+        store
+            .persist_trajectory(TursoTrajectoryInsert {
+                tenant: &entry.tenant,
+                entity_type: &entry.entity_type,
+                entity_id: &entry.entity_id,
+                action: &entry.action,
+                success: entry.success,
+                from_status: entry.from_status.as_deref(),
+                to_status: entry.to_status.as_deref(),
+                error: entry.error.as_deref(),
+                agent_id: entry.agent_id.as_deref(),
+                session_id: entry.session_id.as_deref(),
+                authz_denied: entry.authz_denied,
+                denied_resource: entry.denied_resource.as_deref(),
+                denied_module: entry.denied_module.as_deref(),
+                source,
+                spec_governed: entry.spec_governed,
+                created_at: &entry.timestamp,
+                request_body: request_body_json.as_deref(),
+                intent: entry.intent.as_deref(),
+                matched_policy_ids: matched_policy_ids_json.as_deref(),
+            })
+            .await
+            .map_err(|e| {
+                format!(
+                    "failed to persist trajectory entry for {}/{}/{} action {} in turso-routed: {e}",
+                    entry.tenant, entry.entity_type, entry.entity_id, entry.action
+                )
+            })
+    }
+}
