+++ crates/temper-server/src/storage/mod.rs
@@ -18,19 +18,18 @@ use std::time::Duration;
 
 use sqlx::PgPool;
 use temper_runtime::persistence::{EventStore, PersistenceEnvelope, PersistenceError};
-use temper_store_postgres::{PostgresEventStore, PostgresPolicyRow, PostgresTrajectoryInsert};
+use temper_store_postgres::{PostgresEventStore, PostgresPolicyRow};
 use temper_store_turso::{
-    ActionStats, AgentSummary, DesignTimeEventRow, EvolutionRecordRow, FeatureRequestRow,
-    OtsTrajectoryParams, OtsTrajectoryRow, PolicyDenialPatternRow, PolicyRow as TursoPolicyRow,
-    TenantStoreRouter, TenantUserRow, TursoEventStore, TursoTrajectoryInsert, TursoTrajectoryRow,
-    TursoWasmInvocationInsert, TursoWasmInvocationRow, TursoWasmModuleMetadataRow,
-    UnmetIntentAggRow, store::TrajectoryStats,
+    AgentSummary, DesignTimeEventRow, EvolutionRecordRow, FeatureRequestRow, OtsTrajectoryParams,
+    OtsTrajectoryRow, PolicyDenialPatternRow, PolicyRow as TursoPolicyRow, TenantStoreRouter,
+    TenantUserRow, TursoEventStore, TursoTrajectoryRow, TursoWasmInvocationInsert,
+    TursoWasmInvocationRow, TursoWasmModuleMetadataRow, UnmetIntentAggRow, store::TrajectoryStats,
 };
 
 use crate::platform_store::PlatformStore;
 #[cfg(feature = "sim")]
 use crate::platform_store::SimPlatformStore;
-use crate::state::trajectory::{TrajectoryEntry, TrajectorySource};
+use crate::state::trajectory::TrajectoryEntry;
 
 mod published_artifacts;
 pub use published_artifacts::{
@@ -875,1748 +874,6 @@ impl StorageStack {
     }
 }
 
-struct SingleMetadataStoreProvider {
-    store: Arc<dyn MetadataStore>,
-}
-
-impl SingleMetadataStoreProvider {
-    fn new<T>(store: Arc<T>) -> Self
-    where
-        T: MetadataStore + 'static,
-    {
-        Self { store }
-    }
-}
-
-#[async_trait::async_trait]
-impl MetadataStoreProvider for SingleMetadataStoreProvider {
-    fn platform_store(&self) -> Option<Arc<dyn MetadataStore>> {
-        Some(self.store.clone())
-    }
-
-    async fn store_for_tenant(&self, _tenant: &str) -> Option<Arc<dyn MetadataStore>> {
-        Some(self.store.clone())
-    }
-
-    async fn all_stores(&self) -> Vec<Arc<dyn MetadataStore>> {
-        vec![self.store.clone()]
-    }
-}
-
-struct SingleTursoStoreProvider {
-    store: Arc<TursoEventStore>,
-}
-
-impl SingleTursoStoreProvider {
-    fn new(store: Arc<TursoEventStore>) -> Self {
-        Self { store }
-    }
-}
-
-fn tenant_admin_unsupported() -> PersistenceError {
-    PersistenceError::Storage("tenant management requires routed Turso storage".to_string())
-}
-
-#[async_trait::async_trait]
-impl TursoStoreProvider for SingleTursoStoreProvider {
-    fn supports_tenant_admin(&self) -> bool {
-        false
-    }
-
-    fn platform_store(&self) -> Option<TursoEventStore> {
-        Some(self.store.as_ref().clone())
-    }
-
-    async fn store_for_tenant(&self, _tenant: &str) -> Option<TursoEventStore> {
-        Some(self.store.as_ref().clone())
-    }
-
-    async fn all_stores(&self) -> Vec<TursoEventStore> {
-        vec![self.store.as_ref().clone()]
-    }
-
-    async fn connected_tenants(&self) -> Vec<String> {
-        Vec::new()
-    }
-
-    async fn tenants_for_user(
-        &self,
-        _user_id: &str,
-    ) -> Result<Vec<TenantUserRow>, PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-
-    async fn register_tenant(&self, _tenant_id: &str) -> Result<TursoEventStore, PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-
-    async fn list_tenants(&self) -> Result<Vec<String>, PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-
-    async fn remove_tenant(&self, _tenant_id: &str) -> Result<bool, PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-
-    async fn add_tenant_user(
-        &self,
-        _tenant_id: &str,
-        _user_id: &str,
-        _role: &str,
-    ) -> Result<(), PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-
-    async fn list_tenant_users(
-        &self,
-        _tenant_id: &str,
-    ) -> Result<Vec<TenantUserRow>, PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-
-    async fn remove_tenant_user(
-        &self,
-        _tenant_id: &str,
-        _user_id: &str,
-    ) -> Result<(), PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-
-    async fn ensure_tenant(&self, _tenant_id: &str) -> Result<bool, PersistenceError> {
-        Err(tenant_admin_unsupported())
-    }
-}
-
-struct TenantRoutedMetadataStoreProvider {
-    router: TenantStoreRouter,
-}
-
-impl TenantRoutedMetadataStoreProvider {
-    fn new(router: TenantStoreRouter) -> Self {
-        Self { router }
-    }
-}
-
-#[async_trait::async_trait]
-impl MetadataStoreProvider for TenantRoutedMetadataStoreProvider {
-    fn platform_store(&self) -> Option<Arc<dyn MetadataStore>> {
-        Some(Arc::new(self.router.platform_store().clone()) as Arc<dyn MetadataStore>)
-    }
-
-    async fn store_for_tenant(&self, tenant: &str) -> Option<Arc<dyn MetadataStore>> {
-        self.router
-            .store_for_tenant(tenant)
-            .await
-            .ok()
-            .map(|store| Arc::new(store) as Arc<dyn MetadataStore>)
-    }
-
-    async fn all_stores(&self) -> Vec<Arc<dyn MetadataStore>> {
-        let mut stores =
-            vec![Arc::new(self.router.platform_store().clone()) as Arc<dyn MetadataStore>];
-        for tenant_id in self.router.connected_tenants().await {
-            if let Ok(store) = self.router.store_for_tenant(&tenant_id).await {
-                stores.push(Arc::new(store) as Arc<dyn MetadataStore>);
-            }
-        }
-        stores
-    }
-}
-
-struct TenantRoutedTursoStoreProvider {
-    router: TenantStoreRouter,
-}
-
-impl TenantRoutedTursoStoreProvider {
-    fn new(router: TenantStoreRouter) -> Self {
-        Self { router }
-    }
-}
-
-#[async_trait::async_trait]
-impl TursoStoreProvider for TenantRoutedTursoStoreProvider {
-    fn supports_tenant_admin(&self) -> bool {
-        true
-    }
-
-    fn platform_store(&self) -> Option<TursoEventStore> {
-        Some(self.router.platform_store().clone())
-    }
-
-    async fn store_for_tenant(&self, tenant: &str) -> Option<TursoEventStore> {
-        self.router.store_for_tenant(tenant).await.ok()
-    }
-
-    async fn all_stores(&self) -> Vec<TursoEventStore> {
-        let mut stores = vec![self.router.platform_store().clone()];
-        for tenant_id in self.router.connected_tenants().await {
-            if let Ok(store) = self.router.store_for_tenant(&tenant_id).await {
-                stores.push(store);
-            }
-        }
-        stores
-    }
-
-    async fn connected_tenants(&self) -> Vec<String> {
-        self.router.connected_tenants().await
-    }
-
-    async fn tenants_for_user(
-        &self,
-        user_id: &str,
-    ) -> Result<Vec<TenantUserRow>, PersistenceError> {
-        self.router.tenants_for_user(user_id).await
-    }
-
-    async fn register_tenant(&self, tenant_id: &str) -> Result<TursoEventStore, PersistenceError> {
-        self.router.register_tenant(tenant_id).await
-    }
-
-    async fn list_tenants(&self) -> Result<Vec<String>, PersistenceError> {
-        self.router.list_tenants().await
-    }
-
-    async fn remove_tenant(&self, tenant_id: &str) -> Result<bool, PersistenceError> {
-        self.router.remove_tenant(tenant_id).await
-    }
-
-    async fn add_tenant_user(
-        &self,
-        tenant_id: &str,
-        user_id: &str,
-        role: &str,
-    ) -> Result<(), PersistenceError> {
-        self.router.add_tenant_user(tenant_id, user_id, role).await
-    }
-
-    async fn list_tenant_users(
-        &self,
-        tenant_id: &str,
-    ) -> Result<Vec<TenantUserRow>, PersistenceError> {
-        self.router.list_tenant_users(tenant_id).await
-    }
-
-    async fn remove_tenant_user(
-        &self,
-        tenant_id: &str,
-        user_id: &str,
-    ) -> Result<(), PersistenceError> {
-        self.router.remove_tenant_user(tenant_id, user_id).await
-    }
-
-    async fn ensure_tenant(&self, tenant_id: &str) -> Result<bool, PersistenceError> {
-        self.router.ensure_tenant(tenant_id).await
-    }
-}
-
-#[async_trait::async_trait]
-impl PolicyStore for PostgresEventStore {
-    async fn save_policy(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        cedar_text: &str,
-        created_by: &str,
-    ) -> Result<bool, String> {
-        self.save_policy(tenant, policy_id, cedar_text, created_by)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn load_policies_for_tenant(&self, tenant: &str) -> Result<Vec<PolicyStoreRow>, String> {
-        self.load_policies_for_tenant(tenant)
-            .await
-            .map(|rows| rows.into_iter().map(PolicyStoreRow::from).collect())
-            .map_err(|e| e.to_string())
-    }
-
-    async fn load_all_policies(&self) -> Result<Vec<PolicyStoreRow>, String> {
-        self.load_all_policies()
-            .await
-            .map(|rows| rows.into_iter().map(PolicyStoreRow::from).collect())
-            .map_err(|e| e.to_string())
-    }
-
-    async fn toggle_policy_enabled(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        enabled: bool,
-    ) -> Result<bool, String> {
-        self.toggle_policy_enabled(tenant, policy_id, enabled)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn update_policy_text(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        cedar_text: &str,
-        created_by: &str,
-    ) -> Result<bool, String> {
-        self.update_policy_text(tenant, policy_id, cedar_text, created_by)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn delete_policy(&self, tenant: &str, policy_id: &str) -> Result<(), String> {
-        self.delete_policy(tenant, policy_id)
-            .await
-            .map_err(|e| e.to_string())
-    }
-}
-
-#[async_trait::async_trait]
-impl PolicyStore for TursoEventStore {
-    async fn save_policy(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        cedar_text: &str,
-        created_by: &str,
-    ) -> Result<bool, String> {
-        self.save_policy(tenant, policy_id, cedar_text, created_by)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn load_policies_for_tenant(&self, tenant: &str) -> Result<Vec<PolicyStoreRow>, String> {
-        self.load_policies_for_tenant(tenant)
-            .await
-            .map(|rows| rows.into_iter().map(PolicyStoreRow::from).collect())
-            .map_err(|e| e.to_string())
-    }
-
-    async fn load_all_policies(&self) -> Result<Vec<PolicyStoreRow>, String> {
-        self.load_all_policies()
-            .await
-            .map(|rows| rows.into_iter().map(PolicyStoreRow::from).collect())
-            .map_err(|e| e.to_string())
-    }
-
-    async fn toggle_policy_enabled(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        enabled: bool,
-    ) -> Result<bool, String> {
-        self.toggle_policy_enabled(tenant, policy_id, enabled)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn update_policy_text(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        cedar_text: &str,
-        created_by: &str,
-    ) -> Result<bool, String> {
-        self.update_policy_text(tenant, policy_id, cedar_text, created_by)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn delete_policy(&self, tenant: &str, policy_id: &str) -> Result<(), String> {
-        self.delete_policy(tenant, policy_id)
-            .await
-            .map_err(|e| e.to_string())
-    }
-}
-
-#[async_trait::async_trait]
-impl PolicyStore for TenantStoreRouter {
-    async fn save_policy(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        cedar_text: &str,
-        created_by: &str,
-    ) -> Result<bool, String> {
-        let store = self
-            .store_for_tenant(tenant)
-            .await
-            .map_err(|e| e.to_string())?;
-        store
-            .save_policy(tenant, policy_id, cedar_text, created_by)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn load_policies_for_tenant(&self, tenant: &str) -> Result<Vec<PolicyStoreRow>, String> {
-        let store = self
-            .store_for_tenant(tenant)
-            .await
-            .map_err(|e| e.to_string())?;
-        store
-            .load_policies_for_tenant(tenant)
-            .await
-            .map(|rows| rows.into_iter().map(PolicyStoreRow::from).collect())
-            .map_err(|e| e.to_string())
-    }
-
-    async fn load_all_policies(&self) -> Result<Vec<PolicyStoreRow>, String> {
-        let mut rows: Vec<PolicyStoreRow> = self
-            .platform_store()
-            .load_all_policies()
-            .await
-            .map(|rows| rows.into_iter().map(PolicyStoreRow::from).collect())
-            .map_err(|e| e.to_string())?;
-        for tenant_id in self.connected_tenants().await {
-            if let Ok(store) = self.store_for_tenant(&tenant_id).await {
-                let mut tenant_rows: Vec<PolicyStoreRow> = store
-                    .load_all_policies()
-                    .await
-                    .map(|rows| rows.into_iter().map(PolicyStoreRow::from).collect())
-                    .map_err(|e| e.to_string())?;
-                rows.append(&mut tenant_rows);
-            }
-        }
-        Ok(rows)
-    }
-
-    async fn toggle_policy_enabled(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        enabled: bool,
-    ) -> Result<bool, String> {
-        let store = self
-            .store_for_tenant(tenant)
-            .await
-            .map_err(|e| e.to_string())?;
-        store
-            .toggle_policy_enabled(tenant, policy_id, enabled)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn update_policy_text(
-        &self,
-        tenant: &str,
-        policy_id: &str,
-        cedar_text: &str,
-        created_by: &str,
-    ) -> Result<bool, String> {
-        let store = self
-            .store_for_tenant(tenant)
-            .await
-            .map_err(|e| e.to_string())?;
-        store
-            .update_policy_text(tenant, policy_id, cedar_text, created_by)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    async fn delete_policy(&self, tenant: &str, policy_id: &str) -> Result<(), String> {
-        let store = self
-            .store_for_tenant(tenant)
-            .await
-            .map_err(|e| e.to_string())?;
-        store
-            .delete_policy(tenant, policy_id)
-            .await
-            .map_err(|e| e.to_string())
-    }
-}
-
-impl BackendNamedStore for PostgresEventStore {
-    fn backend_name(&self) -> &'static str {
-        "postgres"
-    }
-}
-
-impl BackendNamedStore for TursoEventStore {
-    fn backend_name(&self) -> &'static str {
-        "turso"
-    }
-}
-
-#[async_trait::async_trait]
-impl ObserveReadStore for PostgresEventStore {
-    async fn load_recent_trajectories(
-        &self,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.load_recent_trajectories(limit)
-            .await
-            .map(|rows| rows.into_iter().map(pg_trajectory_to_turso).collect())
-    }
-
-    async fn load_unmet_intent_rows(&self) -> Result<Vec<UnmetIntentAggRow>, PersistenceError> {
-        self.load_unmet_intent_rows()
-            .await
-            .map(|rows| rows.into_iter().map(pg_unmet_to_turso).collect())
-    }
-
-    async fn load_submit_spec_timestamps(
-        &self,
-    ) -> Result<BTreeMap<String, String>, PersistenceError> {
-        self.load_submit_spec_timestamps().await
-    }
-
-    async fn count_trajectories_by_tenant(
-        &self,
-    ) -> Result<BTreeMap<String, u64>, PersistenceError> {
-        self.count_trajectories_by_tenant().await
-    }
-
-    async fn query_trajectory_stats(
-        &self,
-        entity_type: Option<&str>,
-        action: Option<&str>,
-        success_filter: Option<bool>,
-        failed_limit: i64,
-    ) -> Result<TrajectoryStats, PersistenceError> {
-        self.query_trajectory_stats(entity_type, action, success_filter, failed_limit)
-            .await
-            .map(pg_stats_to_turso)
-    }
-
-    async fn query_trajectories_by_agent(
-        &self,
-        agent_id: &str,
-        tenant: Option<&str>,
-        entity_type: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.query_trajectories_by_agent(agent_id, tenant, entity_type, limit)
-            .await
-            .map(|rows| rows.into_iter().map(pg_trajectory_to_turso).collect())
-    }
-
-    async fn query_agent_summaries(
-        &self,
-        tenant: Option<&str>,
-    ) -> Result<Vec<AgentSummary>, PersistenceError> {
-        self.query_agent_summaries(tenant)
-            .await
-            .map(|rows| rows.into_iter().map(pg_agent_summary_to_turso).collect())
-    }
-}
-
-#[async_trait::async_trait]
-impl ObserveReadStore for TursoEventStore {
-    async fn load_recent_trajectories(
-        &self,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.load_recent_trajectories(limit).await
-    }
-
-    async fn load_unmet_intent_rows(&self) -> Result<Vec<UnmetIntentAggRow>, PersistenceError> {
-        self.load_unmet_intent_rows().await
-    }
-
-    async fn load_submit_spec_timestamps(
-        &self,
-    ) -> Result<BTreeMap<String, String>, PersistenceError> {
-        self.load_submit_spec_timestamps().await
-    }
-
-    async fn count_trajectories_by_tenant(
-        &self,
-    ) -> Result<BTreeMap<String, u64>, PersistenceError> {
-        self.count_trajectories_by_tenant().await
-    }
-
-    async fn query_trajectory_stats(
-        &self,
-        entity_type: Option<&str>,
-        action: Option<&str>,
-        success_filter: Option<bool>,
-        failed_limit: i64,
-    ) -> Result<TrajectoryStats, PersistenceError> {
-        self.query_trajectory_stats(entity_type, action, success_filter, failed_limit)
-            .await
-    }
-
-    async fn query_trajectories_by_agent(
-        &self,
-        agent_id: &str,
-        tenant: Option<&str>,
-        entity_type: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.query_trajectories_by_agent(agent_id, tenant, entity_type, limit)
-            .await
-    }
-
-    async fn query_agent_summaries(
-        &self,
-        tenant: Option<&str>,
-    ) -> Result<Vec<AgentSummary>, PersistenceError> {
-        self.query_agent_summaries(tenant).await
-    }
-}
-
-#[async_trait::async_trait]
-impl EvolutionStore for PostgresEventStore {
-    async fn upsert_feature_request(
-        &self,
-        id: &str,
-        category: &str,
-        description: &str,
-        frequency: i64,
-        trajectory_refs_json: &str,
-        disposition: &str,
-        developer_notes: Option<&str>,
-    ) -> Result<(), PersistenceError> {
-        self.upsert_feature_request(
-            id,
-            category,
-            description,
-            frequency,
-            trajectory_refs_json,
-            disposition,
-            developer_notes,
-        )
-        .await
-    }
-
-    async fn list_feature_requests(
-        &self,
-        disposition: Option<&str>,
-    ) -> Result<Vec<FeatureRequestRow>, PersistenceError> {
-        self.list_feature_requests(disposition)
-            .await
-            .map(|rows| rows.into_iter().map(pg_feature_request_to_turso).collect())
-    }
-
-    async fn update_feature_request(
-        &self,
-        id: &str,
-        disposition: &str,
-        developer_notes: Option<&str>,
-    ) -> Result<bool, PersistenceError> {
-        self.update_feature_request(id, disposition, developer_notes)
-            .await
-    }
-
-    async fn insert_evolution_record(
-        &self,
-        id: &str,
-        record_type: &str,
-        status: &str,
-        created_by: &str,
-        derived_from: Option<&str>,
-        data_json: &str,
-    ) -> Result<(), PersistenceError> {
-        self.insert_evolution_record(id, record_type, status, created_by, derived_from, data_json)
-            .await
-    }
-
-    async fn get_evolution_record(
-        &self,
-        id: &str,
-    ) -> Result<Option<EvolutionRecordRow>, PersistenceError> {
-        self.get_evolution_record(id)
-            .await
-            .map(|row| row.map(pg_evolution_record_to_turso))
-    }
-
-    async fn list_evolution_records(
-        &self,
-        record_type: Option<&str>,
-        status: Option<&str>,
-    ) -> Result<Vec<EvolutionRecordRow>, PersistenceError> {
-        self.list_evolution_records(record_type, status)
-            .await
-            .map(|rows| rows.into_iter().map(pg_evolution_record_to_turso).collect())
-    }
-
-    async fn list_ranked_insights(&self) -> Result<Vec<EvolutionRecordRow>, PersistenceError> {
-        self.list_ranked_insights()
-            .await
-            .map(|rows| rows.into_iter().map(pg_evolution_record_to_turso).collect())
-    }
-}
-
-#[async_trait::async_trait]
-impl EvolutionStore for TursoEventStore {
-    async fn upsert_feature_request(
-        &self,
-        id: &str,
-        category: &str,
-        description: &str,
-        frequency: i64,
-        trajectory_refs_json: &str,
-        disposition: &str,
-        developer_notes: Option<&str>,
-    ) -> Result<(), PersistenceError> {
-        self.upsert_feature_request(
-            id,
-            category,
-            description,
-            frequency,
-            trajectory_refs_json,
-            disposition,
-            developer_notes,
-        )
-        .await
-    }
-
-    async fn list_feature_requests(
-        &self,
-        disposition: Option<&str>,
-    ) -> Result<Vec<FeatureRequestRow>, PersistenceError> {
-        self.list_feature_requests(disposition).await
-    }
-
-    async fn update_feature_request(
-        &self,
-        id: &str,
-        disposition: &str,
-        developer_notes: Option<&str>,
-    ) -> Result<bool, PersistenceError> {
-        self.update_feature_request(id, disposition, developer_notes)
-            .await
-    }
-
-    async fn insert_evolution_record(
-        &self,
-        id: &str,
-        record_type: &str,
-        status: &str,
-        created_by: &str,
-        derived_from: Option<&str>,
-        data_json: &str,
-    ) -> Result<(), PersistenceError> {
-        self.insert_evolution_record(id, record_type, status, created_by, derived_from, data_json)
-            .await
-    }
-
-    async fn get_evolution_record(
-        &self,
-        id: &str,
-    ) -> Result<Option<EvolutionRecordRow>, PersistenceError> {
-        self.get_evolution_record(id).await
-    }
-
-    async fn list_evolution_records(
-        &self,
-        record_type: Option<&str>,
-        status: Option<&str>,
-    ) -> Result<Vec<EvolutionRecordRow>, PersistenceError> {
-        self.list_evolution_records(record_type, status).await
-    }
-
-    async fn list_ranked_insights(&self) -> Result<Vec<EvolutionRecordRow>, PersistenceError> {
-        self.list_ranked_insights().await
-    }
-}
-
-#[async_trait::async_trait]
-impl DesignTimeEventStore for PostgresEventStore {
-    async fn insert_design_time_event(
-        &self,
-        kind: &str,
-        entity_type: &str,
-        tenant: &str,
-        summary: &str,
-        level: Option<&str>,
-        passed: Option<bool>,
-        step_number: Option<i64>,
-        total_steps: Option<i64>,
-    ) -> Result<(), PersistenceError> {
-        self.insert_design_time_event(
-            kind,
-            entity_type,
-            tenant,
-            summary,
-            level,
-            passed,
-            step_number,
-            total_steps,
-        )
-        .await
-    }
-
-    async fn list_design_time_events(
-        &self,
-        tenant: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<DesignTimeEventRow>, PersistenceError> {
-        self.list_design_time_events(tenant, limit)
-            .await
-            .map(|rows| {
-                rows.into_iter()
-                    .map(pg_design_time_event_to_turso)
-                    .collect()
-            })
-    }
-}
-
-#[async_trait::async_trait]
-impl DesignTimeEventStore for TursoEventStore {
-    async fn insert_design_time_event(
-        &self,
-        kind: &str,
-        entity_type: &str,
-        tenant: &str,
-        summary: &str,
-        level: Option<&str>,
-        passed: Option<bool>,
-        step_number: Option<i64>,
-        total_steps: Option<i64>,
-    ) -> Result<(), PersistenceError> {
-        self.insert_design_time_event(
-            kind,
-            entity_type,
-            tenant,
-            summary,
-            level,
-            passed,
-            step_number,
-            total_steps,
-        )
-        .await
-    }
-
-    async fn list_design_time_events(
-        &self,
-        tenant: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<DesignTimeEventRow>, PersistenceError> {
-        self.list_design_time_events(tenant, limit).await
-    }
-}
-
-#[async_trait::async_trait]
-impl OtsStore for PostgresEventStore {
-    async fn persist_ots_trajectory(
-        &self,
-        params: &OtsTrajectoryParams<'_>,
-    ) -> Result<(), PersistenceError> {
-        self.persist_ots_trajectory(&temper_store_postgres::PostgresOtsTrajectoryParams {
-            trajectory_id: params.trajectory_id,
-            tenant: params.tenant,
-            agent_id: params.agent_id,
-            session_id: params.session_id,
-            outcome: params.outcome,
-            turn_count: params.turn_count,
-            data: params.data,
-        })
-        .await
-    }
-
-    async fn list_ots_trajectories(
-        &self,
-        tenant: &str,
-        agent_id: Option<&str>,
-        outcome: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<OtsTrajectoryRow>, PersistenceError> {
-        self.list_ots_trajectories(tenant, agent_id, outcome, limit)
-            .await
-            .map(|rows| rows.into_iter().map(pg_ots_to_turso).collect())
-    }
-
-    async fn get_ots_trajectory(
-        &self,
-        trajectory_id: &str,
-    ) -> Result<Option<String>, PersistenceError> {
-        self.get_ots_trajectory(trajectory_id).await
-    }
-}
-
-#[async_trait::async_trait]
-impl OtsStore for TursoEventStore {
-    async fn persist_ots_trajectory(
-        &self,
-        params: &OtsTrajectoryParams<'_>,
-    ) -> Result<(), PersistenceError> {
-        self.persist_ots_trajectory(params).await
-    }
-
-    async fn list_ots_trajectories(
-        &self,
-        tenant: &str,
-        agent_id: Option<&str>,
-        outcome: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<OtsTrajectoryRow>, PersistenceError> {
-        self.list_ots_trajectories(tenant, agent_id, outcome, limit)
-            .await
-    }
-
-    async fn get_ots_trajectory(
-        &self,
-        trajectory_id: &str,
-    ) -> Result<Option<String>, PersistenceError> {
-        self.get_ots_trajectory(trajectory_id).await
-    }
-}
-
-#[async_trait::async_trait]
-impl BlobStore for PostgresEventStore {
-    async fn put_blob(&self, key: &str, data: &[u8]) -> Result<(), String> {
-        self.put_blob(key, data).await
-    }
-
-    async fn put_blob_with_ttl(
-        &self,
-        key: &str,
-        data: &[u8],
-        ttl: Option<Duration>,
-    ) -> Result<(), String> {
-        self.put_blob_with_ttl(key, data, ttl).await
-    }
-
-    async fn sweep_expired_blobs(&self, max_rows: u64) -> Result<u64, String> {
-        self.sweep_expired_blobs(max_rows).await
-    }
-
-    async fn get_blob(&self, key: &str) -> Result<Option<Vec<u8>>, String> {
-        self.get_blob(key).await
-    }
-}
-
-#[async_trait::async_trait]
-impl BlobStore for TursoEventStore {
-    async fn put_blob(&self, key: &str, data: &[u8]) -> Result<(), String> {
-        self.put_blob(key, data).await
-    }
-
-    async fn put_blob_with_ttl(
-        &self,
-        key: &str,
-        data: &[u8],
-        ttl: Option<Duration>,
-    ) -> Result<(), String> {
-        self.put_blob_with_ttl(key, data, ttl).await
-    }
-
-    async fn sweep_expired_blobs(&self, max_rows: u64) -> Result<u64, String> {
-        self.sweep_expired_blobs(max_rows).await
-    }
-
-    async fn get_blob(&self, key: &str) -> Result<Option<Vec<u8>>, String> {
-        self.get_blob(key).await
-    }
-}
-
-#[async_trait::async_trait]
-impl AuthzAnalyticsStore for PostgresEventStore {
-    async fn upsert_policy_denial_pattern(
-        &self,
-        tenant: &str,
-        agent_type: Option<&str>,
-        action: &str,
-        resource_type: &str,
-        resource_id: &str,
-        timestamp: &str,
-    ) -> Result<(), PersistenceError> {
-        self.upsert_policy_denial_pattern(
-            tenant,
-            agent_type,
-            action,
-            resource_type,
-            resource_id,
-            timestamp,
-        )
-        .await
-    }
-
-    async fn load_policy_denial_patterns(
-        &self,
-        tenant: &str,
-    ) -> Result<Vec<PolicyDenialPatternRow>, PersistenceError> {
-        self.load_policy_denial_patterns(tenant)
-            .await
-            .map(|rows| rows.into_iter().map(pg_denial_pattern_to_turso).collect())
-    }
-}
-
-#[async_trait::async_trait]
-impl AuthzAnalyticsStore for TursoEventStore {
-    async fn upsert_policy_denial_pattern(
-        &self,
-        tenant: &str,
-        agent_type: Option<&str>,
-        action: &str,
-        resource_type: &str,
-        resource_id: &str,
-        timestamp: &str,
-    ) -> Result<(), PersistenceError> {
-        self.upsert_policy_denial_pattern(
-            tenant,
-            agent_type,
-            action,
-            resource_type,
-            resource_id,
-            timestamp,
-        )
-        .await
-    }
-
-    async fn load_policy_denial_patterns(
-        &self,
-        tenant: &str,
-    ) -> Result<Vec<PolicyDenialPatternRow>, PersistenceError> {
-        self.load_policy_denial_patterns(tenant).await
-    }
-}
-
-#[async_trait::async_trait]
-impl DecisionStore for PostgresEventStore {
-    async fn query_decisions(
-        &self,
-        tenant: &str,
-        status: Option<&str>,
-    ) -> Result<Vec<String>, PersistenceError> {
-        self.query_decisions(tenant, status).await
-    }
-
-    async fn query_all_decisions(
-        &self,
-        status: Option<&str>,
-    ) -> Result<Vec<String>, PersistenceError> {
-        self.query_all_decisions(status).await
-    }
-
-    async fn get_pending_decision(&self, id: &str) -> Result<Option<String>, PersistenceError> {
-        self.get_pending_decision(id).await
-    }
-}
-
-#[async_trait::async_trait]
-impl DecisionStore for TursoEventStore {
-    async fn query_decisions(
-        &self,
-        tenant: &str,
-        status: Option<&str>,
-    ) -> Result<Vec<String>, PersistenceError> {
-        self.query_decisions(tenant, status).await
-    }
-
-    async fn query_all_decisions(
-        &self,
-        status: Option<&str>,
-    ) -> Result<Vec<String>, PersistenceError> {
-        self.query_all_decisions(status).await
-    }
-
-    async fn get_pending_decision(&self, id: &str) -> Result<Option<String>, PersistenceError> {
-        self.get_pending_decision(id).await
-    }
-}
-
-#[async_trait::async_trait]
-impl WasmMetadataStore for PostgresEventStore {
-    async fn load_wasm_module_metadata_all_tenants(
-        &self,
-    ) -> Result<Vec<TursoWasmModuleMetadataRow>, PersistenceError> {
-        self.load_wasm_module_metadata_all_tenants()
-            .await
-            .map(|rows| rows.into_iter().map(pg_wasm_metadata_to_turso).collect())
-    }
-
-    async fn delete_wasm_module(
-        &self,
-        tenant: &str,
-        module_name: &str,
-    ) -> Result<bool, PersistenceError> {
-        self.delete_wasm_module(tenant, module_name).await
-    }
-}
-
-#[async_trait::async_trait]
-impl WasmMetadataStore for TursoEventStore {
-    async fn load_wasm_module_metadata_all_tenants(
-        &self,
-    ) -> Result<Vec<TursoWasmModuleMetadataRow>, PersistenceError> {
-        self.load_wasm_module_metadata_all_tenants().await
-    }
-
-    async fn delete_wasm_module(
-        &self,
-        tenant: &str,
-        module_name: &str,
-    ) -> Result<bool, PersistenceError> {
-        self.delete_wasm_module(tenant, module_name).await
-    }
-}
-
-#[async_trait::async_trait]
-impl WasmInvocationStore for PostgresEventStore {
-    async fn persist_wasm_invocation(
-        &self,
-        entry: &TursoWasmInvocationInsert<'_>,
-    ) -> Result<(), PersistenceError> {
-        self.persist_wasm_invocation(&temper_store_postgres::PostgresWasmInvocationInsert {
-            tenant: entry.tenant,
-            entity_type: entry.entity_type,
-            entity_id: entry.entity_id,
-            module_name: entry.module_name,
-            trigger_action: entry.trigger_action,
-            callback_action: entry.callback_action,
-            success: entry.success,
-            error: entry.error,
-            duration_ms: entry.duration_ms,
-            created_at: entry.created_at,
-        })
-        .await
-    }
-
-    async fn load_recent_wasm_invocations(
-        &self,
-        limit: i64,
-    ) -> Result<Vec<TursoWasmInvocationRow>, PersistenceError> {
-        self.load_recent_wasm_invocations(limit)
-            .await
-            .map(|rows| rows.into_iter().map(pg_wasm_invocation_to_turso).collect())
-    }
-}
-
-#[async_trait::async_trait]
-impl WasmInvocationStore for TursoEventStore {
-    async fn persist_wasm_invocation(
-        &self,
-        entry: &TursoWasmInvocationInsert<'_>,
-    ) -> Result<(), PersistenceError> {
-        self.persist_wasm_invocation(entry).await
-    }
-
-    async fn load_recent_wasm_invocations(
-        &self,
-        limit: i64,
-    ) -> Result<Vec<TursoWasmInvocationRow>, PersistenceError> {
-        self.load_recent_wasm_invocations(limit).await
-    }
-}
-
-fn pg_trajectory_to_turso(row: temper_store_postgres::PostgresTrajectoryRow) -> TursoTrajectoryRow {
-    TursoTrajectoryRow {
-        tenant: row.tenant,
-        entity_type: row.entity_type,
-        entity_id: row.entity_id,
-        action: row.action,
-        success: row.success,
-        from_status: row.from_status,
-        to_status: row.to_status,
-        error: row.error,
-        agent_id: row.agent_id,
-        session_id: row.session_id,
-        authz_denied: row.authz_denied,
-        denied_resource: row.denied_resource,
-        denied_module: row.denied_module,
-        source: row.source,
-        spec_governed: row.spec_governed,
-        created_at: row.created_at,
-        request_body: row.request_body,
-        intent: row.intent,
-        matched_policy_ids: row.matched_policy_ids,
-    }
-}
-
-fn pg_unmet_to_turso(row: temper_store_postgres::PostgresUnmetIntentAggRow) -> UnmetIntentAggRow {
-    UnmetIntentAggRow {
-        entity_type: row.entity_type,
-        action: row.action,
-        error: row.error,
-        count: row.count,
-        first_seen: row.first_seen,
-        last_seen: row.last_seen,
-    }
-}
-
-fn pg_stats_to_turso(stats: temper_store_postgres::PostgresTrajectoryStats) -> TrajectoryStats {
-    TrajectoryStats {
-        total: stats.total,
-        success_count: stats.success_count,
-        error_count: stats.error_count,
-        success_rate: stats.success_rate,
-        by_action: stats
-            .by_action
-            .into_iter()
-            .map(|(name, action)| {
-                (
-                    name,
-                    ActionStats {
-                        total: action.total,
-                        success: action.success,
-                        error: action.error,
-                    },
-                )
-            })
-            .collect(),
-        failed_intents: stats
-            .failed_intents
-            .into_iter()
-            .map(pg_trajectory_to_turso)
-            .collect(),
-    }
-}
-
-fn pg_agent_summary_to_turso(row: temper_store_postgres::PostgresAgentSummary) -> AgentSummary {
-    AgentSummary {
-        agent_id: row.agent_id,
-        total_actions: row.total_actions,
-        success_count: row.success_count,
-        error_count: row.error_count,
-        denial_count: row.denial_count,
-        success_rate: row.success_rate,
-        last_active_at: row.last_active_at,
-    }
-}
-
-fn pg_feature_request_to_turso(
-    row: temper_store_postgres::PostgresFeatureRequestRow,
-) -> FeatureRequestRow {
-    FeatureRequestRow {
-        id: row.id,
-        category: row.category,
-        description: row.description,
-        frequency: row.frequency,
-        trajectory_refs: row.trajectory_refs,
-        disposition: row.disposition,
-        developer_notes: row.developer_notes,
-        created_at: row.created_at,
-        updated_at: row.updated_at,
-    }
-}
-
-fn pg_evolution_record_to_turso(
-    row: temper_store_postgres::PostgresEvolutionRecordRow,
-) -> EvolutionRecordRow {
-    EvolutionRecordRow {
-        id: row.id,
-        record_type: row.record_type,
-        status: row.status,
-        created_by: row.created_by,
-        derived_from: row.derived_from,
-        data: row.data,
-        timestamp: row.timestamp,
-    }
-}
-
-fn pg_design_time_event_to_turso(
-    row: temper_store_postgres::PostgresDesignTimeEventRow,
-) -> DesignTimeEventRow {
-    DesignTimeEventRow {
-        id: row.id,
-        kind: row.kind,
-        entity_type: row.entity_type,
-        tenant: row.tenant,
-        summary: row.summary,
-        level: row.level,
-        passed: row.passed,
-        step_number: row.step_number,
-        total_steps: row.total_steps,
-        created_at: row.created_at,
-    }
-}
-
-fn pg_ots_to_turso(row: temper_store_postgres::PostgresOtsTrajectoryRow) -> OtsTrajectoryRow {
-    OtsTrajectoryRow {
-        trajectory_id: row.trajectory_id,
-        tenant: row.tenant,
-        agent_id: row.agent_id,
-        session_id: row.session_id,
-        outcome: row.outcome,
-        turn_count: row.turn_count,
-        created_at: row.created_at,
-    }
-}
-
-fn pg_denial_pattern_to_turso(
-    row: temper_store_postgres::PostgresPolicyDenialPatternRow,
-) -> PolicyDenialPatternRow {
-    PolicyDenialPatternRow {
-        tenant: row.tenant,
-        agent_type: row.agent_type,
-        action: row.action,
-        resource_type: row.resource_type,
-        count: row.count,
-        first_seen: row.first_seen,
-        last_seen: row.last_seen,
-        distinct_resource_ids_json: row.distinct_resource_ids_json,
-    }
-}
-
-fn pg_wasm_metadata_to_turso(
-    row: temper_store_postgres::PostgresWasmModuleMetadataRow,
-) -> TursoWasmModuleMetadataRow {
-    TursoWasmModuleMetadataRow {
-        tenant: row.tenant,
-        module_name: row.module_name,
-        sha256_hash: row.sha256_hash,
-        size_bytes: row.size_bytes,
-        updated_at: row.updated_at,
-    }
-}
-
-fn pg_wasm_invocation_to_turso(
-    row: temper_store_postgres::PostgresWasmInvocationRow,
-) -> TursoWasmInvocationRow {
-    TursoWasmInvocationRow {
-        tenant: row.tenant,
-        entity_type: row.entity_type,
-        entity_id: row.entity_id,
-        module_name: row.module_name,
-        trigger_action: row.trigger_action,
-        callback_action: row.callback_action,
-        success: row.success,
-        error: row.error,
-        duration_ms: row.duration_ms,
-        created_at: row.created_at,
-    }
-}
-
-#[async_trait::async_trait]
-impl QueryPlaneStore for PostgresEventStore {
-    async fn upsert_projection(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_id: &str,
-        status: &str,
-        fields: &serde_json::Value,
-        sequence_nr: u64,
-    ) -> Result<(), PersistenceError> {
-        self.upsert_query_projection(tenant, entity_type, entity_id, status, fields, sequence_nr)
-            .await
-    }
-
-    async fn remove_projection(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_id: &str,
-    ) -> Result<(), PersistenceError> {
-        self.remove_query_projection(tenant, entity_type, entity_id)
-            .await
-    }
-
-    async fn query_field_index(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        where_clause: &str,
-        params: Vec<String>,
-    ) -> Result<Option<Vec<String>>, PersistenceError> {
-        PostgresEventStore::query_field_index(self, tenant, entity_type, where_clause, params)
-            .await
-            .map(Some)
-    }
-
-    async fn load_projection_fields_many(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_ids: &[String],
-        field_names: &[&str],
-    ) -> Result<Option<Vec<QueryProjectionFieldsRow>>, PersistenceError> {
-        self.load_query_projection_fields_many(tenant, entity_type, entity_ids, field_names)
-            .await
-            .map(|rows| {
-                Some(
-                    rows.into_iter()
-                        .map(|row| QueryProjectionFieldsRow {
-                            entity_id: row.entity_id,
-                            status: row.status,
-                            fields: row.fields,
-                        })
-                        .collect(),
-                )
-            })
-    }
-
-    async fn load_entity_catalog_rows(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_ids: &[String],
-    ) -> Result<Option<Vec<EntityCatalogRow>>, PersistenceError> {
-        self.load_entity_catalog_rows_pg(tenant, entity_type, entity_ids)
-            .await
-            .map(|rows| {
-                Some(
-                    rows.into_iter()
-                        .map(|row| EntityCatalogRow {
-                            entity_id: row.entity_id,
-                            status: row.status,
-                            fields: row.fields,
-                            sequence_nr: row.sequence_nr,
-                        })
-                        .collect(),
-                )
-            })
-    }
-
-    async fn projected_entity_counts_by_tenant(
-        &self,
-    ) -> Result<Option<Vec<(String, u64)>>, PersistenceError> {
-        PostgresEventStore::projected_entity_counts_by_tenant(self)
-            .await
-            .map(Some)
-    }
-}
-
-#[async_trait::async_trait]
-impl QueryPlaneStore for TursoEventStore {
-    async fn upsert_projection(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_id: &str,
-        status: &str,
-        fields: &serde_json::Value,
-        sequence_nr: u64,
-    ) -> Result<(), PersistenceError> {
-        self.upsert_query_projection(tenant, entity_type, entity_id, status, fields, sequence_nr)
-            .await
-    }
-
-    async fn remove_projection(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_id: &str,
-    ) -> Result<(), PersistenceError> {
-        self.remove_query_projection(tenant, entity_type, entity_id)
-            .await
-    }
-
-    async fn query_field_index(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        where_clause: &str,
-        params: Vec<String>,
-    ) -> Result<Option<Vec<String>>, PersistenceError> {
-        TursoEventStore::query_field_index(self, tenant, entity_type, where_clause, params)
-            .await
-            .map(Some)
-    }
-
-    async fn load_projection_fields_many(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_ids: &[String],
-        field_names: &[&str],
-    ) -> Result<Option<Vec<QueryProjectionFieldsRow>>, PersistenceError> {
-        self.load_query_projection_fields_many(tenant, entity_type, entity_ids, field_names)
-            .await
-            .map(|rows| {
-                Some(
-                    rows.into_iter()
-                        .map(|row| QueryProjectionFieldsRow {
-                            entity_id: row.entity_id,
-                            status: row.status,
-                            fields: row.fields,
-                        })
-                        .collect(),
-                )
-            })
-    }
-
-    async fn load_entity_catalog_rows(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_ids: &[String],
-    ) -> Result<Option<Vec<EntityCatalogRow>>, PersistenceError> {
-        TursoEventStore::load_entity_catalog_rows(self, tenant, entity_type, entity_ids)
-            .await
-            .map(|rows| {
-                Some(
-                    rows.into_iter()
-                        .map(|row| EntityCatalogRow {
-                            entity_id: row.entity_id,
-                            status: row.status,
-                            fields: row.fields,
-                            sequence_nr: row.sequence_nr,
-                        })
-                        .collect(),
-                )
-            })
-    }
-
-    async fn projected_entity_counts_by_tenant(
-        &self,
-    ) -> Result<Option<Vec<(String, u64)>>, PersistenceError> {
-        TursoEventStore::projected_entity_counts_by_tenant(self)
-            .await
-            .map(Some)
-    }
-}
-
-#[async_trait::async_trait]
-impl DataOnlyCreateStore for PostgresEventStore {
-    async fn create_data_only_entity(
-        &self,
-        record: DataOnlyCreateRecord<'_>,
-    ) -> Result<u64, PersistenceError> {
-        self.create_data_only_entity_native(
-            record.tenant,
-            record.entity_type,
-            record.entity_id,
-            record.status,
-            record.fields,
-            record.event,
-        )
-        .await
-    }
-}
-
-#[async_trait::async_trait]
-impl QueryPlaneStore for TenantStoreRouter {
-    async fn upsert_projection(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_id: &str,
-        status: &str,
-        fields: &serde_json::Value,
-        sequence_nr: u64,
-    ) -> Result<(), PersistenceError> {
-        let store = self.store_for_tenant(tenant).await?;
-        store
-            .upsert_query_projection(tenant, entity_type, entity_id, status, fields, sequence_nr)
-            .await
-    }
-
-    async fn remove_projection(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_id: &str,
-    ) -> Result<(), PersistenceError> {
-        let store = self.store_for_tenant(tenant).await?;
-        store
-            .remove_query_projection(tenant, entity_type, entity_id)
-            .await
-    }
-
-    async fn query_field_index(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        where_clause: &str,
-        params: Vec<String>,
-    ) -> Result<Option<Vec<String>>, PersistenceError> {
-        let store = self.store_for_tenant(tenant).await?;
-        TursoEventStore::query_field_index(&store, tenant, entity_type, where_clause, params)
-            .await
-            .map(Some)
-    }
-
-    async fn load_projection_fields_many(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_ids: &[String],
-        field_names: &[&str],
-    ) -> Result<Option<Vec<QueryProjectionFieldsRow>>, PersistenceError> {
-        let store = self.store_for_tenant(tenant).await?;
-        store
-            .load_query_projection_fields_many(tenant, entity_type, entity_ids, field_names)
-            .await
-            .map(|rows| {
-                Some(
-                    rows.into_iter()
-                        .map(|row| QueryProjectionFieldsRow {
-                            entity_id: row.entity_id,
-                            status: row.status,
-                            fields: row.fields,
-                        })
-                        .collect(),
-                )
-            })
-    }
-
-    async fn load_entity_catalog_rows(
-        &self,
-        tenant: &str,
-        entity_type: &str,
-        entity_ids: &[String],
-    ) -> Result<Option<Vec<EntityCatalogRow>>, PersistenceError> {
-        let store = self.store_for_tenant(tenant).await?;
-        TursoEventStore::load_entity_catalog_rows(&store, tenant, entity_type, entity_ids)
-            .await
-            .map(|rows| {
-                Some(
-                    rows.into_iter()
-                        .map(|row| EntityCatalogRow {
-                            entity_id: row.entity_id,
-                            status: row.status,
-                            fields: row.fields,
-                            sequence_nr: row.sequence_nr,
-                        })
-                        .collect(),
-                )
-            })
-    }
-
-    async fn projected_entity_counts_by_tenant(
-        &self,
-    ) -> Result<Option<Vec<(String, u64)>>, PersistenceError> {
-        let mut counts = Vec::new();
-        for tenant_id in self.connected_tenants().await {
-            let store = self.store_for_tenant(&tenant_id).await?;
-            if let Some((_, count)) = TursoEventStore::projected_entity_counts_by_tenant(&store)
-                .await?
-                .into_iter()
-                .find(|(tenant, _)| tenant == &tenant_id)
-            {
-                counts.push((tenant_id, count));
-            }
-        }
-        Ok(Some(counts))
-    }
-}
-
-fn trajectory_source_label(source: &TrajectorySource) -> &'static str {
-    match source {
-        TrajectorySource::Entity => "Entity",
-        TrajectorySource::Platform => "Platform",
-        TrajectorySource::Authz => "Authz",
-    }
-}
-
-fn trajectory_request_body_json(entry: &TrajectoryEntry) -> Option<String> {
-    entry.request_body.as_ref().and_then(|value| {
-        let serialized = serde_json::to_string(value).ok()?;
-        Some(if serialized.len() > 4096 {
-            let mut end = 4096;
-            while !serialized.is_char_boundary(end) {
-                end -= 1;
-            }
-            serialized[..end].to_string()
-        } else {
-            serialized
-        })
-    })
-}
-
-fn trajectory_matched_policy_ids_json(entry: &TrajectoryEntry) -> Option<String> {
-    entry
-        .matched_policy_ids
-        .as_ref()
-        .and_then(|ids| serde_json::to_string(ids).ok())
-}
-
-#[async_trait::async_trait]
-impl TrajectorySink for PostgresEventStore {
-    async fn persist_trajectory_entry(&self, entry: &TrajectoryEntry) -> Result<(), String> {
-        let matched_policy_ids_json = trajectory_matched_policy_ids_json(entry);
-        let request_body_json = trajectory_request_body_json(entry);
-        let source = entry.source.as_ref().map(trajectory_source_label);
-
-        self.persist_trajectory(PostgresTrajectoryInsert {
-            tenant: &entry.tenant,
-            entity_type: &entry.entity_type,
-            entity_id: &entry.entity_id,
-            action: &entry.action,
-            success: entry.success,
-            from_status: entry.from_status.as_deref(),
-            to_status: entry.to_status.as_deref(),
-            error: entry.error.as_deref(),
-            agent_id: entry.agent_id.as_deref(),
-            session_id: entry.session_id.as_deref(),
-            authz_denied: entry.authz_denied,
-            denied_resource: entry.denied_resource.as_deref(),
-            denied_module: entry.denied_module.as_deref(),
-            source,
-            spec_governed: entry.spec_governed,
-            created_at: &entry.timestamp,
-            request_body: request_body_json.as_deref(),
-            intent: entry.intent.as_deref(),
-            matched_policy_ids: matched_policy_ids_json.as_deref(),
-        })
-        .await
-        .map_err(|e| {
-            format!(
-                "failed to persist trajectory entry for {}/{}/{} action {} in postgres: {e}",
-                entry.tenant, entry.entity_type, entry.entity_id, entry.action
-            )
-        })
-    }
-}
-
-#[async_trait::async_trait]
-impl TrajectorySink for TursoEventStore {
-    async fn persist_trajectory_entry(&self, entry: &TrajectoryEntry) -> Result<(), String> {
-        let matched_policy_ids_json = trajectory_matched_policy_ids_json(entry);
-        let request_body_json = trajectory_request_body_json(entry);
-        let source = entry.source.as_ref().map(trajectory_source_label);
-
-        self.persist_trajectory(TursoTrajectoryInsert {
-            tenant: &entry.tenant,
-            entity_type: &entry.entity_type,
-            entity_id: &entry.entity_id,
-            action: &entry.action,
-            success: entry.success,
-            from_status: entry.from_status.as_deref(),
-            to_status: entry.to_status.as_deref(),
-            error: entry.error.as_deref(),
-            agent_id: entry.agent_id.as_deref(),
-            session_id: entry.session_id.as_deref(),
-            authz_denied: entry.authz_denied,
-            denied_resource: entry.denied_resource.as_deref(),
-            denied_module: entry.denied_module.as_deref(),
-            source,
-            spec_governed: entry.spec_governed,
-            created_at: &entry.timestamp,
-            request_body: request_body_json.as_deref(),
-            intent: entry.intent.as_deref(),
-            matched_policy_ids: matched_policy_ids_json.as_deref(),
-        })
-        .await
-        .map_err(|e| {
-            format!(
-                "failed to persist trajectory entry for {}/{}/{} action {} in turso: {e}",
-                entry.tenant, entry.entity_type, entry.entity_id, entry.action
-            )
-        })
-    }
-}
-
-#[async_trait::async_trait]
-impl TrajectorySink for TenantStoreRouter {
-    async fn persist_trajectory_entry(&self, entry: &TrajectoryEntry) -> Result<(), String> {
-        let store = self.store_for_tenant(&entry.tenant).await.map_err(|e| {
-            format!(
-                "failed to resolve tenant store for trajectory entry {}/{}/{} action {}: {e}",
-                entry.tenant, entry.entity_type, entry.entity_id, entry.action
-            )
-        })?;
-        let matched_policy_ids_json = trajectory_matched_policy_ids_json(entry);
-        let request_body_json = trajectory_request_body_json(entry);
-        let source = entry.source.as_ref().map(trajectory_source_label);
-
-        store
-            .persist_trajectory(TursoTrajectoryInsert {
-                tenant: &entry.tenant,
-                entity_type: &entry.entity_type,
-                entity_id: &entry.entity_id,
-                action: &entry.action,
-                success: entry.success,
-                from_status: entry.from_status.as_deref(),
-                to_status: entry.to_status.as_deref(),
-                error: entry.error.as_deref(),
-                agent_id: entry.agent_id.as_deref(),
-                session_id: entry.session_id.as_deref(),
-                authz_denied: entry.authz_denied,
-                denied_resource: entry.denied_resource.as_deref(),
-                denied_module: entry.denied_module.as_deref(),
-                source,
-                spec_governed: entry.spec_governed,
-                created_at: &entry.timestamp,
-                request_body: request_body_json.as_deref(),
-                intent: entry.intent.as_deref(),
-                matched_policy_ids: matched_policy_ids_json.as_deref(),
-            })
-            .await
-            .map_err(|e| {
-                format!(
-                    "failed to persist trajectory entry for {}/{}/{} action {} in turso-routed: {e}",
-                    entry.tenant, entry.entity_type, entry.entity_id, entry.action
-                )
-            })
-    }
-}
+mod backends;
+mod providers;
+use providers::*;
